fix(tests): clear bots cache
This works similar as with anonymous user, so clear it manually.
